// packages/editor/novel-writer/src/lib/types.ts
export type WorkflowStage
  = | 'requirements'
    | 'worldbuilding'
    | 'worldbuilding_review'
    | 'characters'
    | 'characters_review'
    | 'story_background'
    | 'story_background_review'
    | 'outline'
    | 'outline_review'
    | 'scene_writing'
    | 'chapter_editing'
    | 'chapter_supervision'
    | 'supervision'
    | 'editing'
    | 'final'